% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_pmf_R6.R
\docType{class}
\name{BABipartiteKernel}
\alias{BABipartiteKernel}
\title{BA_BipartiteKernel R6 Class}
\description{
Implements a Barabási–Albert (BA) style kernel for bipartite dynamic networks, inheriting from MarkKernel.
}
\seealso{
\link{MarkKernel}, \link{BAKernel}
}
\section{Super class}{
\code{\link[hawkesGrowthNet:MarkKernel]{hawkesGrowthNet::MarkKernel}} -> \code{BABipartiteKernel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BABipartiteKernel-clone}{\code{BABipartiteKernel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="as_legacy_fun"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-as_legacy_fun'><code>hawkesGrowthNet::MarkKernel$as_legacy_fun()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="compute"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-compute'><code>hawkesGrowthNet::MarkKernel$compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="initialize"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-initialize'><code>hawkesGrowthNet::MarkKernel$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="pmf"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-pmf'><code>hawkesGrowthNet::MarkKernel$pmf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="sample"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-sample'><code>hawkesGrowthNet::MarkKernel$sample()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BABipartiteKernel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BABipartiteKernel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BABipartiteKernel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
