% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn_ig}
\alias{bn_ig}
\title{Internal function, takes a
bipartite network and makes it into a bipartite igraph
mainly useful for plotting}
\usage{
bn_ig(net)
}
\description{
Internal function, takes a
bipartite network and makes it into a bipartite igraph
mainly useful for plotting
}
