% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CollegeMsg}
\alias{CollegeMsg}
\title{CollegeMsg}
\format{
Each newline represents an edge
}
\source{
\url{https://snap.stanford.edu/data/CollegeMsg.html}
}
\usage{
data(CollegeMsg)
}
\description{
College messaging temporal network from \url{https://snap.stanford.edu/data/CollegeMsg.html}
}
\examples{
data(CollegeMsg)
head(CollegeMsg)
}
\keyword{dataset}
