% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet_inhom.R
\name{compute_kde_background}
\alias{compute_kde_background}
\title{Compute KDE Background}
\usage{
compute_kde_background(times, windowT, bw = "nrd0")
}
\arguments{
\item{times}{Numeric vector of event times.}

\item{windowT}{Vector c(min, max) for the time window.}

\item{bw}{Bandwidth for density (default "nrd0").}
}
\value{
A list containing:
\item{mu_func}{Approxfun function to get background density at any t}
\item{mu_vec}{Vector of background densities evaluated at 'times'}
\item{integral}{The integral of the density over the window (approx 1)}
}
\description{
Computes a smoothed background rate shape using Kernel Density Estimation.
}
