% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_events_from_net}
\alias{create_events_from_net}
\title{Title}
\usage{
create_events_from_net(
  net,
  node_time_attr = "time",
  edge_time_attr = "time",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Preserves network attributes. Defaults to expecting "time" as edge / vertex
attribute for event / birth time.
}
