% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_net_from_events}
\alias{create_net_from_events}
\title{Title}
\usage{
create_net_from_events(
  events,
  net_init_fun = net_init,
  net0 = NULL,
  net_step_fun = net_add_event,
  debug = FALSE
)
}
\arguments{
\item{net_step_fun}{}
}
\description{
Title
}
