% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_pmf_R6.R
\docType{class}
\name{CSKernel}
\alias{CSKernel}
\title{CSKernel R6 Class}
\format{
An \code{R6Class} object inheriting from \code{MarkKernel}.
}
\description{
Implements a Change-Statistic (CS) kernel for HawkesNet, inheriting from MarkKernel. Uses ERNM-style change statistics to compute edge probabilities. Node growth is controlled by a Poisson process or cutoff logic. Designed for models where edge probabilities depend on network statistics.
}
\section{Methods}{

\describe{
\item{initialize}{Create a new CSKernel object.}
\item{pmf}{Evaluate only: returns log pmf etc. (no sampling).}
\item{sample}{Sample only: returns sampled mark & its log density (no provided mark density).}
\item{compute}{Evaluate and (optionally) sample in one go (back-compat API).}
\item{as_legacy_fun}{Legacy wrapper to match old PMF_mark_* function signatures.}
}
}

\seealso{
\link{MarkKernel}, \link{BAKernel}
}
\section{Super class}{
\code{\link[hawkesGrowthNet:MarkKernel]{hawkesGrowthNet::MarkKernel}} -> \code{CSKernel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CSKernel-clone}{\code{CSKernel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="as_legacy_fun"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-as_legacy_fun'><code>hawkesGrowthNet::MarkKernel$as_legacy_fun()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="compute"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-compute'><code>hawkesGrowthNet::MarkKernel$compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="initialize"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-initialize'><code>hawkesGrowthNet::MarkKernel$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="pmf"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-pmf'><code>hawkesGrowthNet::MarkKernel$pmf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hawkesGrowthNet" data-topic="MarkKernel" data-id="sample"><a href='../../hawkesGrowthNet/html/MarkKernel.html#method-MarkKernel-sample'><code>hawkesGrowthNet::MarkKernel$sample()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CSKernel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CSKernel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSKernel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
