% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edge_ids_where}
\alias{edge_ids_where}
\title{edge_ids_where}
\usage{
edge_ids_where(net, attr, condition)
}
\arguments{
\item{net}{network object}

\item{attr}{edge attribute to apply condition to}

\item{condition}{function on values returns vector of Bools}
}
\value{
return EIDs of the edges that meet tehe condition - for safe feeding into delete.edges
}
\description{
helper function to make sure we delete the right edges when a network has been
subject to deletion of vertices and edges
}
