% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_PMF.R
\name{expected_params_PMF_mark_BA}
\alias{expected_params_PMF_mark_BA}
\title{Expected parameter names for PMF_mark_BA}
\usage{
expected_params_PMF_mark_BA()
}
\value{
List with \code{required} (character vector of parameter names).
}
\description{
Expected parameter names for PMF_mark_BA
}
