% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filtration_to_net}
\alias{filtration_to_net}
\title{FUNCTION_TITLE}
\usage{
filtration_to_net(net, t, equals = FALSE)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}

\item{equals}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
