% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges_per_event.R
\name{get_edges_per_event}
\alias{get_edges_per_event}
\title{Calculate the number of new edges added at each event time}
\usage{
get_edges_per_event(nw)
}
\arguments{
\item{nw}{A network object with a "time" edge attribute}
}
\value{
A data.frame with columns 'time' and 'n_edges'
}
\description{
Calculate the number of new edges added at each event time
}
\examples{
net <- network::network(matrix(c(1,2, 1,3, 2,3), ncol = 2, byrow = TRUE),
                        directed = FALSE)
network::set.edge.attribute(net, "time", c(0.1, 0.1, 0.5))
get_edges_per_event(net)
}
