% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing infix operator}
\usage{
a %||% b
}
\arguments{
\item{a}{An object, possibly \code{NULL}.}

\item{b}{A fallback value to return if \code{a} is \code{NULL}.}
}
\value{
\code{a} if not \code{NULL}, otherwise \code{b}.
}
\description{
Returns the left-hand side if it is not \code{NULL}, otherwise returns the right-hand side.
Useful for providing default values in expressions.
}
\examples{
x <- NULL
y <- 10
x \%||\% y   # returns 10

z <- 5
z \%||\% y   # returns 5
}
