% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ht09_contact_list}
\alias{ht09_contact_list}
\title{Hypertext 2009 contact list (temporal edges)}
\format{
A tab-separated text file: \code{time node_i node_j}.
Time is in seconds; node IDs identify individuals.
}
\source{
Hypertext 2009 conference data; see \code{inst/hypertext_conference/} for usage.
}
\description{
Contact list / proximity network with temporal edge list: time and node pair.
Stored as a tab-separated file in the package \code{data/} directory.
Used by the hypertext-conference example to fit or illustrate Hawkes network growth.
}
\seealso{
\code{\link{CollegeMsg}}, \code{inst/hypertext_conference/hypertext_conference.R}
}
