% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{ks_test_pval_hawkesGrowthNet}
\alias{ks_test_pval_hawkesGrowthNet}
\title{Kolmogorov-Smirnov test p-value for time rescaling (Hawkes growth model)}
\usage{
ks_test_pval_hawkesGrowthNet(params, time_window, mark_filtration)
}
\arguments{
\item{params}{List of parameters.}

\item{time_window}{Numeric \code{c(t0, t1)}.}

\item{mark_filtration}{Observed network.}
}
\value{
P-value of the KS test under the null that rescaled times are uniform.
}
\description{
Kolmogorov-Smirnov test p-value for time rescaling (Hawkes growth model)
}
