% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Title}
\usage{
loglik(
  events,
  params,
  T_end = NULL,
  T0 = 0,
  mark_type = "ba",
  debug = FALSE,
  net_init_fun = net_init,
  net0 = NULL,
  net_step_fun = net_add_event,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
