% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hawkes.R
\name{loglik_temporal_hawk}
\alias{loglik_temporal_hawk}
\title{FUNCTION_TITLE}
\usage{
loglik_temporal_hawk(
  params,
  realiz,
  windowT,
  dists = NULL,
  density_approx = TRUE,
  numeric_integral = FALSE,
  cl = NULL,
  optimized = T
)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{realiz}{PARAM_DESCRIPTION}

\item{windowT}{PARAM_DESCRIPTION}

\item{dists}{PARAM_DESCRIPTION, Default: NULL}

\item{density_approx}{PARAM_DESCRIPTION, Default: TRUE}

\item{numeric_integral}{PARAM_DESCRIPTION, Default: FALSE}

\item{cl}{PARAM_DESCRIPTION, Default: NULL}

\item{optimized}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{clusterApply}}
}
