% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_network_growth_animation}
\alias{make_network_growth_animation}
\title{FUNCTION_TITLE}
\usage{
make_network_growth_animation(net_list, times, adjust = 10, file = NULL)
}
\arguments{
\item{net_list}{PARAM_DESCRIPTION}

\item{times}{PARAM_DESCRIPTION}

\item{adjust}{PARAM_DESCRIPTION, Default: 10}

\item{file}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
