% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{net}
\alias{net}
\title{net}
\format{
A network object of \link{CollegeMsg}
}
\source{
\url{https://snap.stanford.edu/data/CollegeMsg.html}
}
\usage{
data(net)
}
\description{
The network of the College messaging data from \url{https://snap.stanford.edu/data/CollegeMsg.html}
}
\examples{
data(net)
plot(net)
}
\keyword{dataset}
