% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_add_edges.R
\name{net_add_edges}
\alias{net_add_edges}
\title{Title}
\usage{
net_add_edges(net, new_edges, t)
}
\arguments{
\item{new_edges}{data.frame with at least columns "i" and "j". Any additional
columns will be treated as edge attributes.}

\item{t}{}
}
\value{
Updated `network` object.
}
\description{
Title
}
\examples{
# net <- network::network.initialize(0, directed = FALSE)
# net <- net_add_nodes(net, data.frame(id=c("a","b"), stringsAsFactors=FALSE), t=0)
# net <- net_add_edges(net, data.frame(i="a", j="b", w=1), t=0.5)
}
