% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_add_event.R
\name{net_add_event}
\alias{net_add_event}
\title{Add one event to the network (nodes, then edges)}
\usage{
net_add_event(net, new_nodes, new_edges, t)
}
\arguments{
\item{net}{network object}

\item{new_nodes}{data.frame of nodes to add (must have id; may have role)}

\item{new_edges}{data.frame of edges to add (must have i, j; may have attrs)}

\item{t}{numeric scalar time}
}
\value{
updated network object
}
\description{
Add one event to the network (nodes, then edges)
}
