% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_example_sub_component}
\alias{plot_example_sub_component}
\title{Plot example subcomponents of a bipartite igraph
internal function}
\usage{
plot_example_sub_component(
  g,
  size,
  idx = 1,
  cols = c("#E41A1C", "#377EB8"),
  ...
)
}
\description{
Plot example subcomponents of a bipartite igraph
internal function
}
