% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_background.R
\name{plot_kde_background}
\alias{plot_kde_background}
\title{Plot estimated baseline rate and integrated baseline (for sanity checks)}
\usage{
plot_kde_background(mu_fit, rescale = NULL, main = "KDE baseline")
}
\arguments{
\item{mu_fit}{Output from estimate_mu_kde}

\item{rescale}{Optional list with Lambda_grid (for second panel)}

\item{main}{Plot title}
}
\description{
Plot estimated baseline rate and integrated baseline (for sanity checks)
}
