% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{sim_hawkesGrowthNet}
\alias{sim_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
sim_hawkesGrowthNet(
  params,
  time_window,
  PMF_mark,
  cond_intensity,
  hashed_edges = F,
  verbose = F,
  mu_multiplier = 10,
  joint_accept = F,
  n_mark_sample = NULL,
  ...
)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{PMF_mark}{PARAM_DESCRIPTION}

\item{cond_intensity}{PARAM_DESCRIPTION}

\item{hashed_edges}{PARAM_DESCRIPTION, Default: F}

\item{verbose}{PARAM_DESCRIPTION, Default: F}

\item{mu_multiplier}{PARAM_DESCRIPTION, Default: 10}

\item{joint_accept}{PARAM_DESCRIPTION, Default: F}

\item{n_mark_sample}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{as.edgelist}}
\code{\link[hash]{hash}}
}
