% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hawkes_growth.R
\name{sim_hg}
\alias{sim_hg}
\title{Title}
\usage{
sim_hg(
  params,
  T_end,
  T0 = 0,
  state_init_fun = state_init,
  state_step_fun = state_step,
  sim_mark_fun = sim_mark_ba,
  return_ev = TRUE,
  return_state = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
