% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hawkes_growth.R
\name{sim_hg_batched}
\alias{sim_hg_batched}
\title{Title}
\usage{
sim_hg_batched(
  params,
  T_end,
  T0 = 0,
  mu_multiplier = 10,
  state_init_fun = state_init,
  state_step_fun = state_step,
  sim_mark_fun = sim_mark_ba,
  delta = 0.5,
  return_ev = TRUE,
  return_state = FALSE,
  debug = FALSE
)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
