% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_add_edges.R
\name{state_add_edges}
\alias{state_add_edges}
\title{Update deterministic state after an event's edge batch}
\usage{
state_add_edges(state, edges_k, t_k, update_adj = TRUE)
}
\arguments{
\item{state}{list with elements deg, born, adj}

\item{edges_k}{data.frame with columns i, j (may be 0-row)}

\item{t_k}{numeric scalar event time}
}
\value{
updated state
}
\description{
Update deterministic state after an event's edge batch
}
