% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{validate_events}
\alias{validate_events}
\title{Validate an events object}
\usage{
validate_events(events, strict_cols = FALSE)
}
\arguments{
\item{events}{hg_events}

\item{strict_cols}{if TRUE, disallow extra columns beyond the required ones}
}
\value{
Invisibly TRUE; otherwise errors.
}
\description{
Validate an events object
}
