% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hawkes.R
\name{compensator_temporal_hawkes}
\alias{compensator_temporal_hawkes}
\title{FUNCTION_TITLE}
\usage{
compensator_temporal_hawkes(params, realiz, windowT)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{realiz}{PARAM_DESCRIPTION}

\item{windowT}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
