% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{compensators_hawkesGrowthNet}
\alias{compensators_hawkesGrowthNet}
\title{Compensators (integrated intensity) for the Hawkes growth model}
\usage{
compensators_hawkesGrowthNet(params, time_window, mark_filtration)
}
\arguments{
\item{params}{List of parameters.}

\item{time_window}{Numeric \code{c(t0, t1)}.}

\item{mark_filtration}{Observed network.}
}
\value{
Numeric vector of compensator values at each event time.
}
\description{
Compensators (integrated intensity) for the Hawkes growth model
}
