% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesNet.R
\name{compensators_hawkesNet}
\alias{compensators_hawkesNet}
\title{Compensators (integrated intensity) for the Hawkes growth model}
\usage{
compensators_hawkesNet(params, time_window, mark_filtration)
}
\arguments{
\item{params}{List of parameters.}

\item{time_window}{Numeric \code{c(t0, t1)}.}

\item{mark_filtration}{Observed network.}
}
\value{
Numeric vector of compensator values at each event time.
}
\description{
Compensators (integrated intensity) for the Hawkes growth model
}
