% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{cond_intensity}
\alias{cond_intensity}
\title{FUNCTION_TITLE}
\usage{
cond_intensity(
  new_net,
  t,
  mark_filtration = NULL,
  PMF_mark,
  params,
  new_edge_hash = NULL,
  ...
)
}
\arguments{
\item{new_net}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}

\item{mark_filtration}{PARAM_DESCRIPTION, Default: NULL}

\item{PMF_mark}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{new_edge_hash}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
