% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesNet.R
\name{cond_intensity_inhom}
\alias{cond_intensity_inhom}
\title{Conditional intensity at one event time with inhomogeneous background}
\usage{
cond_intensity_inhom(
  new_net,
  t,
  mark_filtration = NULL,
  PMF_mark,
  params,
  mu_at_t,
  new_edge_hash = NULL,
  times = NULL,
  ...
)
}
\arguments{
\item{new_net}{Network state at current time}

\item{t}{Current event time}

\item{mark_filtration}{Full mark filtration}

\item{PMF_mark}{Mark PMF function (e.g. PMF_mark_CS)}

\item{params}{Parameter list (beta_overall, K, CS_params, etc.; mu not used)}

\item{mu_at_t}{Scalar background rate at time t (e.g. from KDE)}

\item{new_edge_hash}{Optional hash of existing edges}

\item{times}{Precomputed times from get_times; if NULL, taken from mark_filtration}

\item{...}{Passed to PMF_mark (e.g. formula_RHS, truncation)}
}
\value{
List with result (intensity), func (function to evaluate intensity at new params), lambda, kernel_sum, decays, diffs
}
\description{
Same as cond_intensity but uses mu_at_t (scalar) instead of params$mu.
}
