% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{events_to_bipartite_net}
\alias{events_to_bipartite_net}
\title{FUNCTION_TITLE}
\usage{
events_to_bipartite_net(
  x,
  pid = "anon_person_id",
  eid = "event_id",
  time = "diff_date"
)
}
\arguments{
\item{x}{A data frame ...}
}
\description{
FUNCTION_DESCRIPTION
}
