% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{events_to_bipartite_net}
\alias{events_to_bipartite_net}
\title{FUNCTION_TITLE}
\usage{
events_to_bipartite_net(events)
}
\arguments{
\item{events}{A named \code{n*3} \code{data.frame} or \code{3} element \code{list} of node interactions with times and
named columns/elements:
\code{i} - \code{numeric} "from" node ID (i.e., node indices),
\code{j} - \code{numeric} "to" node ID (i.e., node indices), &
\code{t} - \code{numeric} time of interaction.}
}
\description{
As \link{events_to_net} but the "from" (\code{i}) and
"to" (\code{j}) nodes are considered to be different types and
only nodes of different type can form an edge. Node IDs can
be \code{character}s.
}
\examples{
\dontrun{
if(interactive()){
events_list <- list(
i = c("A", "B", "C", "D","D"),
j = c(2, 3, 1, 2, 1),
t = c(5, 10, 15, 20, 21))
bip <- events_to_bipartite_net(events_list)
plot(bip, vertex.col = ifelse(get.vertex.attribute(bip, "type") == "type_i", "#E41A1C", "#377EB8"))
 }
}
}
