% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{events_to_net}
\alias{events_to_net}
\title{FUNCTION_TITLE}
\usage{
events_to_net(events_list, net = NULL, directed = FALSE)
}
\arguments{
\item{events_list}{A data frame of interactions
between participants,  with elements \code{i} - from node,
\code{j} - to node, &
\code{t} - time of interaction.}

\item{net}{PARAM_DESCRIPTION, Default: NULL}

\item{directed}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{network}}, \code{\link[network]{attribute.methods}}, \code{\link[network]{add.vertices}}
}
