% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{events_to_net}
\alias{events_to_net}
\title{Convert event list to network}
\usage{
events_to_net(events, net = NULL, directed = FALSE)
}
\arguments{
\item{events}{A named \code{n*3} \code{data.frame} or \code{3} element \code{list} of node interactions with times and
named columns/elements:
\code{i} - \code{numeric} "from" node ID (i.e., node indices),
\code{j} - \code{numeric} "to" node ID (i.e., node indices), &
\code{t} - \code{numeric} time of interaction.}

\item{net}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{directed}{Logical. Default: \code{FALSE}.}
}
\value{
A \code{network} object with time-stamped edges and vertices. The network attribute
\code{n} is set to the largest node index appearing in the events.
}
\description{
Constructs a network from a data frame of event interactions, where each event
specifies a source node, a target node, and a timestamp. Vertices and edges are
added as needed, and the time attribute is set for each.
}
\details{
Each row in \code{events_list} creates (if necessary) the specified nodes and adds an
edge, with time attributes set for both edge and involved vertices.
}
\examples{
\dontrun{
if(interactive()){
events_list <- list(
i = c(1, 2, 3, 1),
j = c(2, 3, 1, 4),
t = c(5, 10, 15, 20))
net <- events_to_net(events_list)
plot(net)
 }
}
}
\seealso{
\code{\link[network]{network}}, \code{\link[network]{attribute.methods}}, \code{\link[network]{add.vertices}}
}
