% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_PMF.R
\name{expected_params_PMF_mark_BA}
\alias{expected_params_PMF_mark_BA}
\title{Expected parameter names for PMF_mark_BA}
\usage{
expected_params_PMF_mark_BA()
}
\value{
List with \code{required} (character vector: \code{beta_edges}, \code{m}) and
\code{optional} (character vector: \code{vertex_categorical}, \code{vertex_categorical_levels}).
}
\description{
Expected parameter names for PMF_mark_BA
}
