% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{fit_hawkesGrowthNet}
\alias{fit_hawkesGrowthNet}
\title{Fit hawkesGrowthNet Model via Maximum Likelihood}
\usage{
fit_hawkesGrowthNet(
  params_init,
  time_window,
  mark_filtration,
  PMF_mark,
  trace = 0,
  REPORT = 10,
  reltol = 1e-08,
  maxit,
  get_hessian = FALSE,
  fixed_params = NULL,
  ...
)
}
\arguments{
\item{params_init}{List. Initial parameter values.}

\item{time_window}{Numeric vector. Time interval for model fitting.}

\item{mark_filtration}{Network. Network history/filtration.}

\item{PMF_mark}{Function. Computes the mark PMF.}

\item{trace}{Integer. Optimizer trace level. Default: 0.}

\item{REPORT}{Integer. Optimizer report frequency. Default: 10.}

\item{reltol}{Numeric. Optimization relative tolerance. Default: 1e-8.}

\item{maxit}{Integer. Maximum number of iterations.}

\item{get_hessian}{Logical. Compute Hessian for SEs? Default: \code{FALSE}.}

\item{fixed_params}{Character vector or NULL. Parameter names to hold fixed. Default: \code{NULL}.}

\item{...}{Additional arguments passed to log-likelihood function.}
}
\value{
List with component:
\item{fit}{Object returned by \code{optim} (MLE fit).}
}
\description{
Fits (estimates parameters for) a hawkesGrowthNet model by maximizing the log-likelihood.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link{loglik_hawkesGrowthNet}}, \code{\link{sim_hawkesGrowthNet}}
}
