% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{fit_hawkesGrowthNet}
\alias{fit_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
fit_hawkesGrowthNet(
  params_init,
  time_window,
  mark_filtration,
  PMF_mark,
  trace = 0,
  REPORT = 10,
  reltol = 1e-08,
  maxit,
  get_hessian = FALSE,
  fixed_params = NULL,
  ...
)
}
\arguments{
\item{params_init}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{PMF_mark}{PARAM_DESCRIPTION}

\item{trace}{PARAM_DESCRIPTION, Default: 0}

\item{maxit}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{events}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
