% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hawkes.R
\name{fit_temporal_hawkes}
\alias{fit_temporal_hawkes}
\title{FUNCTION_TITLE}
\usage{
fit_temporal_hawkes(params_init, realiz, windowT, trace = 0, maxit, ...)
}
\arguments{
\item{params_init}{PARAM_DESCRIPTION}

\item{realiz}{PARAM_DESCRIPTION}

\item{windowT}{PARAM_DESCRIPTION}

\item{trace}{PARAM_DESCRIPTION, Default: 0}

\item{maxit}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
