% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_times}
\alias{get_times}
\title{FUNCTION_TITLE}
\usage{
get_times(net)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
