% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_edge}
\alias{has_edge}
\title{Check whether an edge exists in a hash-backed edge set}
\usage{
has_edge(i, j, edge_hash)
}
\arguments{
\item{i}{Vertex index (head).}

\item{j}{Vertex index (tail).}

\item{edge_hash}{A \code{hash} object (from package \pkg{hash}) with keys of the form \code{"i-j"}.}
}
\value{
Logical: \code{TRUE} if the edge (i, j) is in the hash.
}
\description{
Check whether an edge exists in a hash-backed edge set
}
\examples{
library(hash)
h <- hash("1-2" = TRUE)
has_edge(1, 2, h)
has_edge(1, 3, h)
}
