% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesNet-package.R
\docType{package}
\name{hawkesNet-package}
\alias{hawkesNet}
\alias{hawkesNet-package}
\title{hawkesNet: Simulate and Fit Maximum Likelihood for Hawkesian Network Growth Models}
\description{
Provides tools for simulating and fitting Hawkes-process-driven network growth models via maximum likelihood. Supports both homogeneous and inhomogeneous (KDE-estimated) background rates, Barabási--Albert and change-statistic (ERNM) mark distributions, parallel intensity-cache construction, and goodness-of-fit diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duncan-clark/hawkesNet}
  \item Report bugs at \url{https://github.com/duncan-clark/hawkesNet/issues}
}

}
\author{
\strong{Maintainer}: Duncan Clark \email{dac6@williams.edu} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
