% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{compensators_hawkesGrowthNet}
\alias{compensators_hawkesGrowthNet}
\alias{ks_test_pval_hawkesGrowthNet}
\title{Compensator Process for hawkesGrowthNet}
\usage{
compensators_hawkesGrowthNet(params, time_window, mark_filtration)

ks_test_pval_hawkesGrowthNet(params, time_window, mark_filtration)
}
\arguments{
\item{params}{List. Model parameters.}

\item{time_window}{Numeric vector. Time interval.}

\item{mark_filtration}{Network. Network history/filtration.}
}
\value{
Numeric vector. Compensator values at observed event times.

Numeric. p-value from KS test.
}
\description{
Computes compensator (integrated intensity) values for a hawkesGrowthNet realization.

Performs a KS test on transformed compensator increments to assess model fit.
}
\seealso{
\code{\link{loglik_hawkesGrowthNet}}

\code{\link{compensators_hawkesGrowthNet}}
}
