% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{ks_test_pval_hawkesGrowthNet}
\alias{ks_test_pval_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
ks_test_pval_hawkesGrowthNet(params, time_window, mark_filtration)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{mark_filtration}{PARAM_DESCRIPTION}
}
\description{
FUNCTION_DESCRIPTION
}
