% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hawkes.R
\name{ks_test_pval_temporal}
\alias{ks_test_pval_temporal}
\title{FUNCTION_TITLE}
\usage{
ks_test_pval_temporal(realiz, windowT, hawkes_par)
}
\arguments{
\item{realiz}{PARAM_DESCRIPTION}

\item{windowT}{PARAM_DESCRIPTION}

\item{hawkes_par}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
