% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hawkes.R
\name{ks_test_pval_temporal}
\alias{ks_test_pval_temporal}
\title{KS Test for Temporal Hawkes Process}
\usage{
ks_test_pval_temporal(
  realiz,
  windowT,
  hawkes_par,
  kernel = c("exp", "powerlaw"),
  use_kde = FALSE
)
}
\arguments{
\item{realiz}{list with elements, n,lon,lat,t}

\item{windowT}{vector with elements, start,end}

\item{hawkes_par}{list with elements,   mu,alpha,beta,K}

\item{kernel}{"exp" or "powerlaw"}

\item{use_kde}{Logical/Numeric. Default 0 (FALSE).}
}
\value{
p-value of KS test
}
\description{
Does KS test on pure temporal hawkes
}
