% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R, R/hawkesGrowthNet_inhom.R
\name{loglik_hawkesGrowthNet}
\alias{loglik_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
loglik_hawkesGrowthNet(
  params,
  time_window,
  mark_filtration,
  PMF_mark,
  edge_hash_list = NULL,
  verbose = FALSE,
  do_grad = FALSE,
  intens_funcs = NULL,
  kde_bg = NULL,
  ...
)

loglik_hawkesGrowthNet(
  params,
  time_window,
  mark_filtration,
  PMF_mark,
  edge_hash_list = NULL,
  verbose = FALSE,
  do_grad = FALSE,
  intens_funcs = NULL,
  kde_bg = NULL,
  ...
)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{PMF_mark}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{events}{PARAM_DESCRIPTION}

\item{use_hashing}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{as.edgelist}}
\code{\link[hash]{hash}}
}
