% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{cond_intensity}
\alias{cond_intensity}
\alias{loglik_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
cond_intensity(
  new_net,
  t,
  mark_filtration = NULL,
  PMF_mark,
  params,
  new_edge_hash = NULL,
  ...
)

loglik_hawkesGrowthNet(
  params,
  time_window,
  mark_filtration,
  PMF_mark,
  edge_hash_list = NULL,
  verbose = FALSE,
  do_grad = FALSE,
  ...
)
}
\arguments{
\item{new_net}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}

\item{mark_filtration}{PARAM_DESCRIPTION, Default: NULL}

\item{PMF_mark}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{new_edge_hash}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{edge_hash_list}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{do_grad}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
Computes the conditional intensity function for the \code{hawkesGrowthNet} model
defined as
\deqn{
  \lambda(t,m \mid \mathcal{H}_t) = 
    q(m \mid t,\mathcal{H}_{t}) \left( \lambda(t \mid \mathcal{H}_{t}) \right)
}
where the time decay kernel function, \eqn{g(t_1,t_2 \mid \mathcal{H}_t)},
is given by \eqn{g(t_1,t_2 \mid \mathcal{H}_t) = \exp(-\beta (t_1 - t_2))}.
The mark PMF, \eqn{q(m \mid t,\mathcal{H}_{t})}, is supplied via \link{PMF_mark}.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
data(net, package = "hawkesGrowthNet")
params_ba <-  list(mu = 10,
beta_overall = 0.1,
K = 0.1, beta_edges = 0.1)
mark_filtration <-  network::get.inducedSubgraph(net, v = 1:10)
loglik_ba <- loglik_hawkesGrowthNet(params = params_ba,
                                       time_window = c(0,max(get_times(mark_filtration)$times)),
                                       mark_filtration =  mark_filtration,
                                       PMF_mark = PMF_mark,  
                                       verbose = FALSE)
## CS
require(ernm)
params_cs <-  list(mu = 10,
                     beta_overall = 0.1,
                      K = 0.1,
                      beta_edges = 0.1,
                      node_lambda = 1,
                      CS_params =  c(-10,0,0,0))
loglik_cs <- loglik_hawkesGrowthNet(params = params_cs,
                                       time_window = c(0,max(get_times(mark_filtration)$times)),
                                       mark_filtration =  mark_filtration,
                                       PMF_mark = PMF_mark, type = "CS",
                                       truncation = 1,
                                       formula_RHS = "edges + triangles + star(c(2,3))",
                                       max_node_time = 1,
                                       verbose = FALSE)
 }
}
}
\seealso{
\code{\link{PMF_mark}}

\code{\link[network]{as.edgelist}}
\code{\link[hash]{hash}}
}
