% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_background.R
\name{make_cumhaz_fun}
\alias{make_cumhaz_fun}
\title{Integrated baseline Lambda(t) = integral_0^t mu(s) ds (cumulative hazard)}
\usage{
make_cumhaz_fun(mu_fit)
}
\arguments{
\item{mu_fit}{Output from estimate_mu_kde (list with grid, mu_grid)}
}
\value{
List with Lambda_fun (function t -> Lambda(t)) and Lambda_grid (data.frame t, Lambda)
}
\description{
Integrated baseline Lambda(t) = integral_0^t mu(s) ds (cumulative hazard)
}
\examples{
t <- sort(c(rexp(50, 2), runif(20, 0, 10)))
mu_fit <- estimate_mu_kde(t, windowT = c(0, 10))
ch <- make_cumhaz_fun(mu_fit)
ch$Lambda_fun(5)
}
