% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state.R
\name{make_state}
\alias{make_state}
\title{Construct an initial network state (baseline graph) at observation start T0}
\usage{
make_state(nodes0 = NULL, edges0 = NULL, T0 = 0, implicit_nodes = TRUE)
}
\description{
Construct an initial network state (baseline graph) at observation start T0
}
