% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_add_event.R
\name{net_add_event}
\alias{net_add_event}
\title{Add one event to the network (nodes, then edges)}
\usage{
net_add_event(net, new_nodes, new_edges, t_k)
}
\arguments{
\item{net}{network object, the current network at time t_k}

\item{new_nodes}{data.frame with at least column "id". Any additional
columns will be treated as vertex attributes.}

\item{new_edges}{data.frame with at least columns "i" and "j". Any additional
columns will be treated as edge attributes.}

\item{t_k}{numeric scalar, the event timestamp}
}
\value{
updated network object
}
\description{
Add one event to the network (nodes, then edges)
}
