% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_add_nodes.R
\name{net_add_nodes}
\alias{net_add_nodes}
\title{Title}
\usage{
net_add_nodes(net, new_nodes, t)
}
\arguments{
\item{new_nodes}{data.frame with at least column "node_ids". Any additional
columns will be treated as vertex attributes.}

\item{t}{}
}
\value{
Updated `network` object.x
}
\description{
Title
}
\examples{
# net <- network::network.initialize(0, directed = FALSE)
# net <- net_add_nodes(net, data.frame(id=c("a","b"), stringsAsFactors=FALSE), t=0)
# net <- net_add_edges(net, data.frame(i="a", j="b", w=1), t=0.5)
}
