% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_temporal.R
\name{plot_temporal_diag}
\alias{plot_temporal_diag}
\title{Plot temporal diagnostics from `diag_temporal()`}
\usage{
plot_temporal_diag(
  diag,
  which = c("count", "dt", "lambda", "tau_hist", "tau_qq"),
  ask = FALSE
)
}
\arguments{
\item{diag}{Output of `diag_temporal()`.}

\item{which}{Character vector of plots to show. Any of:
`"count"`, `"dt"`, `"lambda"`, `"tau_hist"`, `"tau_qq"`.}

\item{ask}{Logical; if `TRUE`, prompt between plots.}
}
\description{
Plot temporal diagnostics from `diag_temporal()`
}
