% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_PMF.R
\name{PMF_mark}
\alias{PMF_mark}
\alias{PMF_mark_BA}
\alias{PMF_mark_CS}
\title{Mark probability mass function for the network generation process}
\usage{
PMF_mark(
  time,
  params,
  mark_filtration,
  type = c("BA", "CS"),
  mark = NULL,
  generate_mark = FALSE,
  generate_density = TRUE,
  grad = FALSE,
  new_edge_hash = NULL,
  truncation = NULL,
  formula_RHS,
  mark_decay = "node_entrance",
  model = NULL,
  max_node_time = 10,
  ...
)

PMF_mark_BA(
  time,
  params,
  mark_filtration,
  mark,
  generate_mark,
  generate_density,
  grad,
  new_edge_hash,
  truncation,
  ...
)

PMF_mark_CS(
  time,
  params,
  mark_filtration,
  mark,
  generate_mark,
  generate_density,
  grad,
  new_edge_hash,
  truncation,
  formula_RHS,
  mark_decay,
  model,
  max_node_time,
  ...
)
}
\arguments{
\item{time}{Numeric, the time at which to evaluate the pmf}

\item{params}{A named list of parameter values: for \code{type = "BA"}, numeric values
\code{mu}, \code{beta_overall}, \code{K}, \code{beta_edges}, & \code{node_lambda};
for \code{type = "CS"} additional parameters include \code{CS_params} a vector of coefficients for
\code{formula_RHS}.}

\item{mark_filtration}{PARAM_DESCRIPTION.}

\item{type}{Character, one of \code{"BA"} for Barabási–Albert or
\code{"CS"} for CS Hawkes PMF. Default: \code{"BA"}.}

\item{mark}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{generate_mark}{Logical, Default: \code{FALSE}.}

\item{generate_density}{PARAM_DESCRIPTION, Default: \code{TRUE}.}

\item{grad}{Logical, if \code{TRUE} comupte gradient, Default: \code{FALSE}.}

\item{new_edge_hash}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{truncation}{Default: \code{NULL}.}

\item{formula_RHS}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{mark_decay}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{model}{PARAM_DESCRIPTION, Default: \code{NULL}.}

\item{max_node_time}{Numeric, the last time at which a node can enter the network. Default: \code{10}.}

\item{...}{PARAM_DESCRIPTION, Default: \code{NULL}.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate the mark PMF ...
}
\details{
Computes the mark PMF, \eqn{q(m\vert t,\mathcal{H}_{t})} (see \code{\link{cond_intensity}}).
Currenlty two options: \code{type = "BA"} and \code{type = "CS"}.

For \code{type = "BA"} the Barabasi Albert (BA) preferential attachment model is used where the mark distribution is defined as
\deqn{
  q(m \mid t, \mathcal{H}_t) =
  \prod_{i=1}^{N_{t-}} \left(p_i^{BA}\right)^{e_i} \cdot
  \left(1 - p_i^{BA}\right)^{1 - e_i}.
}
Here, the attachment probability, \eqn{p_i^{BA}}, is defined as
\deqn{
  p_i^{BA} = \frac{\delta_i}{\sum_{k=1}^{N} \delta_k}
}
where \eqn{\delta_{i}^{t} = \exp(\tau \cdot (t - t_i)) \cdot d_{i}^{t}},
and \eqn{d_{i}^{t}} is the degree of node \eqn{i} just before time \eqn{t}.

For \code{type = "CS"} the change statistic (CS) model is used where he mark distribution is defined (similar to above) as
\deqn{
  q(m \mid t, \mathcal{H}_t) =
  \prod_{i=1}^{N_{t-}} \left(p_i^{CS}\right)^{e_i} \cdot
  \left(1 - p_i^{CS}\right)^{1 - e_i}.
}
where the attachment probability, \eqn{p_i^{CS}}, is defined as
\deqn{
p_i^{CS} = \left(\nu + \exp(\tau \cdot(t - t_i))\right)\cdot\frac{1}{1 + \exp(-\theta^{\top} \cdot C_{i,N_t})}
}
}
\examples{
\dontrun{
if(interactive()){
 data(net, package = "hawkesGrowthNet")
time <- get_times(net)$times
## BA
params_ba <-  list(mu = length(time)/max(time), beta_overall = 0.1,K = 0.1,
beta_edges = 0.1,node_lambda = 1)
mark_filtration <-  filtration_to_net(net,10)
pmf_ba <- PMF_mark(time[10],  params_ba, mark_filtration)
## CS
devtools::install_github("duncan-clark/ernm", ref = "R_change_stats")
require(ernm)
params_cs <-  list(mu = length(time)/max(time), beta_overall = 0.1,K = 0.1,
beta_edges = 0.1,node_lambda = 1,CS_params =  c(-10,0,0,0))
pmf_cs <- PMF_mark(time = time[10],  params = params_cs,
mark_filtration = mark_filtration, type = "CS",  truncation = 1,
formula_RHS = "edges + triangles + star(c(2,3))",
max_node_time = 1)
 }
}
}
\references{
Barabási, A.-L. & Albert, R. (1999). Emergence of scaling in random networks. \emph{Science}, 286, 509–512. \doi{10.1126/science.286.5439.509}
}
\seealso{
\code{\link{cond_intensity}}
\code{\link{PMF_mark_BA}}
\code{\link{PMF_mark_CS}}
\code{\link[network]{network}}, \code{\link[network]{add.vertices}}
\code{\link[ernm]{as.BinaryNet}}
}
