% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_PMF.R
\name{PMF_mark_BA}
\alias{PMF_mark_BA}
\title{FUNCTION_TITLE}
\usage{
PMF_mark_BA(
  time,
  params,
  mark_filtration,
  mark = NULL,
  generate_mark = FALSE,
  generate_density = TRUE,
  grad = FALSE,
  new_edge_hash = NULL,
  truncation = NULL
)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{mark_filtration}{PARAM_DESCRIPTION}

\item{mark}{PARAM_DESCRIPTION, Default: NULL}

\item{generate_mark}{PARAM_DESCRIPTION, Default: FALSE}

\item{new_edge_hash}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{network}}, \code{\link[network]{add.vertices}}
}
