% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_PMF.R
\name{PMF_mark_CS}
\alias{PMF_mark_CS}
\title{FUNCTION_TITLE}
\usage{
PMF_mark_CS(
  time,
  params,
  mark_filtration,
  mark = NULL,
  generate_mark = FALSE,
  generate_density = TRUE,
  new_edge_hash = NULL,
  formula_RHS,
  grad = FALSE,
  truncation = 1,
  mark_decay = "node_entrance",
  ...
)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{mark_filtration}{PARAM_DESCRIPTION}

\item{mark}{PARAM_DESCRIPTION, Default: NULL}

\item{generate_mark}{PARAM_DESCRIPTION, Default: FALSE}

\item{new_edge_hash}{PARAM_DESCRIPTION, Default: NULL}

\item{formula_RHS}{PARAM_DESCRIPTION}

\item{grad}{PARAM_DESCRIPTION, Default: FALSE}

\item{truncation}{if truncation = 1, only consider edges from new nodes to old nodes,
truncation = k considers edges from k time steps before the new nodes to the old nodes:}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{network}}, \code{\link[network]{add.vertices}}
\code{\link[ernm]{as.BinaryNet}}
}
