% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomogeneous.R
\name{prepare_inhomogeneous_background}
\alias{prepare_inhomogeneous_background}
\title{Build mu_vec and integral_bg from mark_filtration (network) using KDE}
\usage{
prepare_inhomogeneous_background(
  mark_filtration,
  time_attr = "time",
  bw = NULL,
  grid_n = 2048
)
}
\arguments{
\item{mark_filtration}{Network with vertex and edge time attributes}

\item{time_attr}{Name of the time attribute (default "time")}

\item{bw}{Bandwidth for KDE; NULL = default}

\item{grid_n}{Number of grid points for KDE}
}
\value{
List with mu_vec, integral_bg, times, mu_fit, Lambda_fun
}
\description{
Returns mu_vec aligned with get_times(mark_filtration)$times for use in
loglik_hawkesNet_inhom and fit_hawkesNet_inhom.
}
