% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkesGrowthNet.R
\name{sim_hawkesGrowthNet}
\alias{sim_hawkesGrowthNet}
\title{FUNCTION_TITLE}
\usage{
sim_hawkesGrowthNet(
  params,
  time_window,
  PMF_mark,
  cond_intensity,
  hashed_edges = FALSE,
  verbose = FALSE,
  mu_multiplier = 10,
  joint_accept = FALSE,
  n_mark_sample = NULL,
  ...
)
}
\arguments{
\item{params}{PARAM_DESCRIPTION}

\item{time_window}{PARAM_DESCRIPTION}

\item{PMF_mark}{A function that both generates new mark and calculates the density of existing mark, \link{PMF_mark}.}

\item{cond_intensity}{A function to calculate conditional_intensity, takes in a kernel_func}

\item{hashed_edges}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{mu_multiplier}{PARAM_DESCRIPTION, Default: 10}

\item{joint_accept}{PARAM_DESCRIPTION, Default: FALSE}

\item{n_mark_sample}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{to be passed to \link{PMF_mark}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[network]{as.edgelist}}
\code{\link[hash]{hash}}
}
