% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hawkesNet.R
\name{sim_hawkesNet}
\alias{sim_hawkesNet}
\title{Title}
\usage{
sim_hawkesNet(
  params,
  T_end,
  T0 = 0,
  delta = 0.001,
  mark_type = "ba",
  return_ev = TRUE,
  debug = FALSE,
  net_init_fun = net_init,
  net_step_fun = net_add_event,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
